orginized code
I have orginized the code into different files, next are function calls.
