+++ jpl4py/parser.py
@@ -17,19 +17,19 @@ class jplParser:
         for idx, token in enumerate(self.tokens):
 
             if token.name == "for":
-                print(self.get_nested_content(self.tokens[idx:], hFor))
+                return self.get_nested_content(self.tokens[idx:], hFor)
 
             elif token.name in ["jfn", "cfn", "pfn"]:
-                print(self.get_nested_content(self.tokens[idx:], hFnc))
+                return self.get_nested_content(self.tokens[idx:], hFnc)
 
-    #            elif token.name == "if":
-    #                print(self.get_nested_content(self.tokens[idx:], hIf))
+    #       elif token.name == "if":
+    #            print(self.get_nested_content(self.tokens[idx:], hIf))
 
-    #            elif token.name == "else":
-    #                print(self.get_nested_content(self.tokens[idx:], hElse))
+    #        elif token.name == "else":
+    #            print(self.get_nested_content(self.tokens[idx:], hElse))
 
-    #            elif token.name == "while":
-    #                print(self.get_nested_content(self.tokens[idx:], hWhile))
+    #        elif token.name == "while":
+    #            print(self.get_nested_content(self.tokens[idx:], hWhile))
 
     def get_nested_content(self, tokens: list, cls: object) -> None:
         nest_level: int = 0
